タイプ定義の掃除
