--- packages/kokoas-client/src/types/api.d.ts
@@ -1,6 +1,0 @@
-interface GetRecordParams {
-  app?: number | string,
-  fields?: string[],
-  query?: string,
-  totalCount?: boolean
-}
