+++ packages/kokoas-client/src/types/forms.d.ts
@@ -1,13 +1,2 @@
 
 type TFormStatus = 'busy' | 'disabled' | '';
-
-interface URLParams {
-  projId?: string
-  projEstimateId?: string,
-  custGroupId?: string,
-  menuOpen?: number,
-  invoiceId?: number,
-}
-
-type KeyOfUrlParams = keyof URLParams;
-
