+++ packages/kokoas-client/src/types/types.d.ts
@@ -14,11 +14,3 @@ type OptionNode = {
 };
 
 type Options = Option[];
-
-//type Order = 'asc' | 'desc';
-
-interface EnhancedTableProps<T> {
-  onRequestSort: (event: React.MouseEvent<unknown>, property: T) => void;
-  order: Order;
-  orderBy: string;
-}
